第5次提交 商品模块完成 说明文档未更新

+++ src/main/java/com/warest/mall/service/impl/FileServiceImpl.java
@@ -13,7 +13,7 @@ import java.io.IOException;
 import java.util.UUID;
 
 /**
- * Created by geely
+ * 文件处理用service
  */
 @Service("iFileService")
 public class FileServiceImpl implements IFileService {
@@ -21,29 +21,44 @@ public class FileServiceImpl implements IFileService {
     private Logger logger = LoggerFactory.getLogger(FileServiceImpl.class);
 
 
+    public static void main(String[] args) {
+        String str = "abc.345.def";
+        System.out.println(str.substring(str.lastIndexOf(".")+1));
+    }
+
+    /**
+     * 上传文件
+     * @param file 文件
+     * @param path 上传路径
+     * @return
+     */
     public String upload(MultipartFile file,String path){
-        String fileName = file.getOriginalFilename();
+        String fileName = file.getOriginalFilename(); //文件名
         //扩展名
         //abc.jpg
-        String fileExtensionName = fileName.substring(fileName.lastIndexOf(".")+1);
+        String fileExtensionName = fileName.substring(fileName.lastIndexOf(".")+1); //文件名最后一个点后的内容
         String uploadFileName = UUID.randomUUID().toString()+"."+fileExtensionName;
-        logger.info("开始上传文件,上传文件的文件名:{},上传的路径:{},新文件名:{}",fileName,path,uploadFileName);
+        //todo logger.info("开始上传文件,上传文件的文件名:{},上传的路径:{},新文件名:{}",fileName,path,uploadFileName);
+        logger.info("开始上传文件,上传文件的文件名:{},上传的路径:{},新文件名:{}",new Object[]{fileName,path,uploadFileName});
 
-        File fileDir = new File(path);
+        File fileDir = new File(path); //上传文件夹
         if(!fileDir.exists()){
-            fileDir.setWritable(true);
-            fileDir.mkdirs();
+            fileDir.setWritable(true); //赋予权限可写  保证可以创建
+            fileDir.mkdirs(); //有可能是嵌套文件夹，用mkdirs复数可以一并创建
         }
+        // 创建文件，存放路径和文件名
         File targetFile = new File(path,uploadFileName);
-
 
         try {
             file.transferTo(targetFile);
             //文件已经上传成功了
 
-
-            FTPUtil.uploadFile(Lists.newArrayList(targetFile));
-            //已经上传到ftp服务器上
+            //将文件上传到ftp服务器，可以传多个，但目前只穿了一个？
+            if(!FTPUtil.uploadFile(Lists.newArrayList(targetFile))){
+                logger.error("上传文件失败");
+                return null;
+            }
+            //上传成功，删除网页服务器文件
 
             targetFile.delete();
         } catch (IOException e) {
@@ -52,6 +67,7 @@ public class FileServiceImpl implements IFileService {
         }
         //A:abc.jpg
         //B:abc.jpg
+        //返回目标文件文件名
         return targetFile.getName();
     }
